+++ packages/libs/extension/src/extension/index.ts
@@ -1,0 +1,3 @@
+export * from "./package";
+export * from "./extension";
+export * from "./local-extension";
