+++ packages/libs/extension/src/index.ts
@@ -1,2 +1,4 @@
+export * from "./exceptions";
 export * from "./main";
+export * from "./extension";
 export * from "./system-requirements";
