+++ packages/libs/extension/src/main.ts
@@ -15,14 +15,21 @@ import * as semver from "semver";
 import { Npm } from "./npm";
 import { PackageManager, PackageManagerType } from "./package-manager";
 import { Yarn } from "./yarn";
-import * as fs from "fs";
 import {
   patchPythonPath,
   PythonCommandLine,
-  SystemRequirementError,
   SystemRequirements,
   validateSystemRequirements,
 } from "./system-requirements";
+import { Extension, Package } from "./extension";
+import {
+  UnresolvedPackageException,
+  InvalidPackageIdentityException,
+  ExtensionFolderLocked,
+  PackageInstallationException,
+  MissingStartCommandException,
+  UnsatisfiedSystemRequirementException,
+} from "./exceptions";
 
 function quoteIfNecessary(text: string): string {
   if (text && text.indexOf(" ") > -1 && text.charAt(0) != '"') {
@@ -31,59 +38,6 @@ function quoteIfNecessary(text: string): string {
   return text;
 }
 const nodePath = quoteIfNecessary(process.execPath);
-
-export class UnresolvedPackageException extends Exception {
-  constructor(packageId: string) {
-    super(`Unable to resolve package '${packageId}'.`, 1);
-    Object.setPrototypeOf(this, UnresolvedPackageException.prototype);
-  }
-}
-
-export class InvalidPackageIdentityException extends Exception {
-  constructor(name: string, version: string, message: string) {
-    super(`Package '${name}' - '${version}' is not a valid package reference:\n  ${message}`, 1);
-    Object.setPrototypeOf(this, InvalidPackageIdentityException.prototype);
-  }
-}
-
-export class PackageInstallationException extends Exception {
-  constructor(name: string, version: string, message: string) {
-    super(`Package '${name}' - '${version}' failed to install:\n  ${message}`, 1);
-    Object.setPrototypeOf(this, PackageInstallationException.prototype);
-  }
-}
-
-export class UnsatisfiedEngineException extends Exception {
-  constructor(name: string, version: string, message = "") {
-    super(`Unable to find matching engine '${name}' - '${version} ${message}'`, 1);
-    Object.setPrototypeOf(this, UnsatisfiedEngineException.prototype);
-  }
-}
-
-export class UnsatisfiedSystemRequirementException extends Exception {
-  constructor(extension: Extension, errors: SystemRequirementError[]) {
-    const message = [
-      `System is missing dependencies required by extension '${extension.name}':`,
-      ...errors.map((x) => ` - ${x.name}: ${x.message.replace(/\n/g, "\n ")}`),
-    ].join("\n");
-    super(message, 1);
-    Object.setPrototypeOf(this, UnsatisfiedSystemRequirementException.prototype);
-  }
-}
-
-export class MissingStartCommandException extends Exception {
-  constructor(extension: Extension) {
-    super(`Extension '${extension.id}' is missing the script 'start' in the package.json file`, 1);
-    Object.setPrototypeOf(this, MissingStartCommandException.prototype);
-  }
-}
-
-export class ExtensionFolderLocked extends Exception {
-  constructor(path: string) {
-    super(`Extension Folder '${path}' is locked by another process.`, 1);
-    Object.setPrototypeOf(this, ExtensionFolderLocked.prototype);
-  }
-}
 
 function cmdlineToArray(
   text: string,
@@ -161,142 +115,6 @@ async function getFullPath(command: string, searchPath?: string): Promise<string
   }
 
   return undefined;
-}
-
-/**
- * A Package is a representation of a npm package.
- *
- * Once installed, a Package is an Extension
- */
-export class Package {
-  /* @internal */ public constructor(
-    public resolvedInfo: any,
-    /* @internal */ public packageMetadata: any,
-    /* @internal */ public extensionManager: ExtensionManager,
-  ) {}
-
-  get id(): string {
-    return this.packageMetadata._id;
-  }
-
-  get name(): string {
-    return this.packageMetadata.name;
-  }
-
-  get version(): string {
-    return this.packageMetadata.version;
-  }
-
-  get source(): string {
-    // work around bug that npm doesn't programatically handle exact versions.
-    if (this.resolvedInfo.type === "version" && this.resolvedInfo.registry === true) {
-      return this.packageMetadata._spec + "*";
-    }
-    return this.packageMetadata._spec;
-  }
-
-  async install(force = false): Promise<Extension> {
-    return this.extensionManager.installPackage(this, force);
-  }
-
-  get allVersions(): Promise<Array<string>> {
-    return this.extensionManager.getPackageVersions(this.name);
-  }
-}
-
-/**
- * Extension is an installed Package
- * @extends Package
- * */
-export class Extension extends Package {
-  /* @internal */ public constructor(pkg: Package, private installationPath: string) {
-    super(pkg.resolvedInfo, pkg.packageMetadata, pkg.extensionManager);
-  }
-  /**
-   * The installed location of the package.
-   */
-  public get location(): string {
-    return normalize(`${this.installationPath}/${this.id.replace("/", "_")}`);
-  }
-  /**
-   * The path to the installed npm package (internal to 'location')
-   */
-  public get modulePath(): string {
-    return normalize(`${this.location}/node_modules/${this.name}`);
-  }
-
-  /**
-   * the path to the package.json file for the npm packge.
-   */
-  public get packageJsonPath(): string {
-    return normalize(`${this.modulePath}/package.json`);
-  }
-
-  /**
-   * the path to the readme.md configuration file for the extension.
-   */
-  public get configurationPath(): Promise<string> {
-    return (async () => {
-      const items = await readdir(this.modulePath);
-      for (const each of items) {
-        if (/^readme.md$/i.exec(each)) {
-          const fullPath = normalize(`${this.modulePath}/${each}`);
-          if (await isFile(fullPath)) {
-            return fullPath;
-          }
-        }
-      }
-      return "";
-    })();
-  }
-
-  /** the loaded package.json information */
-  public get definition(): any {
-    const content = fs.readFileSync(this.packageJsonPath).toString();
-    try {
-      return JSON.parse(content);
-    } catch (e) {
-      throw new Error(`Failed to parse package definition at '${this.packageJsonPath}'. ${e}`);
-    }
-  }
-
-  public get configuration(): Promise<string> {
-    return (async () => {
-      const cfgPath = await this.configurationPath;
-      if (cfgPath) {
-        return readFile(cfgPath);
-      }
-      return "";
-    })();
-  }
-
-  async remove(): Promise<void> {
-    return this.extensionManager.removeExtension(this);
-  }
-
-  async start(enableDebugger = false): Promise<ChildProcess> {
-    return this.extensionManager.start(this, enableDebugger);
-  }
-}
-
-/**
- * LocalExtension is a local extension that must not be installed.
- * @extends Extension
- * */
-export class LocalExtension extends Extension {
-  public constructor(pkg: Package, private extensionPath: string) {
-    super(pkg, "");
-  }
-  public get location(): string {
-    return this.extensionPath;
-  }
-  public get modulePath(): string {
-    return this.extensionPath;
-  }
-
-  async remove(): Promise<void> {
-    throw new Error("Cannot remove local extension. Lifetime not our responsibility.");
-  }
 }
 
 async function fetchPackageMetadata(spec: string): Promise<any> {
